// src/common/enums/node-type.enum.ts
export enum NodeType {
  Root = 'root',
  Common = 'common',
  Workspace = 'workspace',
  Application = 'application',
  Dashboard = 'dashboard',
  DashboardGroup = 'dashboardGroup',
  Report = 'report',
  Integration = 'integration',
  IntegrationGroup = 'integrationGroup',
  Workflow = 'workflow',
  ReportGroup = 'reportGroup',
  HTMLCard = 'htmlCard',
  HTMLCardGroup = 'htmlCardGroup',
  UsageStatisticCard = 'usageStatisticCard',
  UsageStatisticCardGroup = 'usageStatisticCardGroup',
  SendEmail = 'sendEmail',
  SaveToFile = 'saveToFile'
}